Worked through CodeBlock#new minefield: major changes to CodeBlock program point methods, some consequences in the NudgeLanguage treetop grammar and helpers, some helpers have appeared in non-codeblock program point classes for tree-walking recursion to work better.
Outcome: when you explicitly invoke CodeBlock.new, you'll get a new CodeBlock with a nil #contents; but when you get myCodeBlock#contents, if it's nil it will launch a parser of its own to fill in the missing material. If you get a CodeBlock from a parser, you'll be set already.

Also, the actual issue (CodeBlock#points not working as specified) is fixed in a similar way.